Bug 492476 - Generic editor support for out of workspace files
Generic Editor can now open files outside the workspace.

If the loaded file in the Generic Editor is opened in another editor
(e.g. the standard text editor) at the same time, too many changed
events are triggered when editing in one editor and then accepting the
changes in the other one.
In the Generic Editor, this results from too many change events from the
command stack. However, it is not clear, yet, why the stack fires that
many events. One idea is that the resource set might trigger changes.

Change-Id: If3470bdd1527d9eae3daac712a006428f00c4548
Signed-off-by: Lucas Koehler <l.koehler93@gmail.com>