Bug 527740 - Local view-model services not inherited by child contexts
Deprecate the APIs that accept a variadic array of view-model services
in construction of contets in favour of new APIs that dynamically
provide services as needed, especially in the creation of child
contexts to inherit service overrides and avoid creation of services
if the child context already exists.

Also add support for Maven test execution on Mac platform.

Change-Id: I46d537ee1498ca8d3c47f8cac80938a4447b0f5b
Signed-off-by: Christian W. Damus <give.a.damus@gmail.com>