Bug 529633 - Workspace compile errors
The AbstractControlSWTRenderer class has a private field of
SWTValidationHelper type and does not expose this type in its API,
yet JDT Compiler still complains that it is required transitively.
To ensure classpath consistency, therefore, the ecp.edit.spi.swt.util
package is added to the uses directive even though it is not strictly
required.

Change-Id: I68bd777bfd49ed39d8943cb1151e016f82f7c879
Signed-off-by: Christian W. Damus <give.a.damus@gmail.com>