Bug 531656 - Use Reference Service to Add a Row in Table Renderer
Use the reference service in the table renderer in case no table service
is registered. This allows to use the template based creation of
elements in a table.

* Added addNewModelElements method with new parameter to
ReferenceService and implementations
* Deprecated old addNewModelElements method.
* Use ReferenceService in TableControlSWTRenderer#addRow

Tests:
* Added tests for the addRow method
* Made the table test bundle a fragment: This was necessary to execute
the protected addRow method
* Refactored some utility methods from SWTTable_PTest out to a new table
test util because these methods are used by other tests in other
bundles, too

Change-Id: I558c0292a36c96ec8965b931d4281cd863a63525
Signed-off-by: Lucas Koehler <lkoehler@eclipsesource.com>