Bug 531723 - Deactivate Checkstyle for Example Projects
Deactivate checkstyle in the project configurations of the make it
happen projects. This is necessary because the make it happen projects
are copied during the build to be then used as a source for the
examples.

This is done by excluding the .checkstyle files when copying the
projects and removing the checkstyle builder from the .project files by
using the maven replacer plugin.

Change-Id: Iac52b2b5e0eb35be1cd7d3dc35deac716c69d95b
Signed-off-by: Lucas Koehler <lkoehler@eclipsesource.com>