Bug 534743 - View Model Editor should allow to create Views for Abstract Classes and Interfaces
* It is now possible to create view models for abstract classes
* These view models can be used in view proxies
* Added a RCPTT test testing the creation and preview of a view for a
simple abstract root EClass (1 attribute and 1 reference)

The Preview can show views for abstract classes with one exception:
It is not possible to add a new object to a reference which has an
EOpposite if the view's root EClass has an instance class (this is the
case if the code for the ecore was generated).
Besides this, the preview works fine for abstract root EClasses.

The Preview checks whether the EClass is abstract or an interface and in
this case creates a dummy wrapper EClass which is used to create a
dynamic EObject.

Further fix:
The priority of the ProxyResolverViewService was lowered. As a
consequence, the service starts earlier. With this, labels and group
names defined in a proxy view can now be correctly resolved by the
LocalizationViewModelService.

Change-Id: I47371bd630722a3ad2e6411194ec6ef05007d718
Signed-off-by: Lucas Koehler <lkoehler@eclipsesource.com>