Bug 537104 - Table with Detail does not handle readonly and enablement correctly
Adapted the Table and Grid detail panel renderers, the table detail
dialog renderer, and the tree master detail renderer to set the detail
view to read-only if the table control is effectively disabled or
read-only.

To achieve propagation of enable/read-only changes to the table control,
the view change listener configured in the AbstractSWTRenderer now
always triggers applyEnable/applyReadonly when the corresponding flag is
changed on any VElement in the renderer VElement's view model.

To still allow navigation of non-control elements (e.g. containers,
categorization), the default behavior of applyEnable in
AbstractSWTRenderer was set to doing nothing.

To allow proper propagation of read-only/enable state, remaining
renderers and the ValidationServiceImpl that were using isReadonly or
isEnabled instead of isEffectivelyReadonly or isEffectivelyEnabled were
changed to use the latter methods.

Change-Id: Id1dc8223572acff58440a1e81230f9f6cc57c03e
Signed-off-by: Lucas Koehler <lkoehler@eclipsesource.com>