Bug 537390 - Add Segments to the DMR [Feature Segment Services]
Adjust existing core services to use segments if they are present in a
DMR and added segment support for feature segments.

* Added interfaces for segment databinding converters, structural change
testers, and domain expanders
* Added feature path implementations for the foregoing interfaces
* Added tests for the implementations
* Extended databinding test model

Adapted the following services to use segments in case a provided DMR
has them. If not, the existing DMR resolvement is used as a fallback:
* EMFFormsDomainExpanderImpl [also provided test cases]
* EMFFormsStructuralChangeTesterImpl [also provided test cases]
* EMFFormsDatabindingImpl [also provided test cases]

Change-Id: I7c34ebc2833ca4df0aea37ce2cca6f08a8153d88
Signed-off-by: Lucas Koehler <lkoehler@eclipsesource.com>