Bug 537390 - Add Segments to the DMR [Hybrid Mode]
Added a view service that allows on-the-fly generation of segments for
DMRs. This view service is NOT used by default but can be activated by
providing the program argument -enableSegmentGeneration

The view service generates the segments by delegating to the new
EMFFormsSegmentGenerator service. The service's implementation itself
delegates the generation to DmrSegmentGenerators which handle the
generation for the concrete DMR sub-types.
In this commit a DmrSegmentGenerator for Feature Path DMRs is provided.

Also added test cases for the view service, The EMFFormsSegmentGenerator
implementation, and the FeaturePathDmrSegmentGenerator.

Change-Id: I1b3f4364599be79798ef7af7a0e40fc0ebe65c06
Signed-off-by: Lucas Koehler <lkoehler@eclipsesource.com>