Bug 542589 - No IItemPropertyDescriptor for feature eModelElement in EClass EAnnotation found.
Removed the control for the eModelElement reference from the EAnnotation
view model. This control is not necessary because the eModelElement
reference is just the opposite to the containment containing the
EAnnotation. If the EAnnotation should be moved, this can simply be done
in the Tree of the EMF Forms ecore editor.

Change-Id: Ib92f3d6cc7c9a7828f1fdbab593b01e8b196719e
Signed-off-by: Lucas Koehler <lkoehler@eclipsesource.com>