Bug 543160 -  [View Model/Validation] Inefficient handling of Diagnostics
On large models, the View Model’s VDiagnostic objects
track a large number of Diagnostics because those
include OK results in addition to problem results.
Optimize the performance of the unique list of
Diagnostics leveraging the fact that we separately
track them in unique hash sets.

Also, it is not necessary to use the unique EList
of the VDiagnostic in the validation service’s
tracking of pending updates, so use simple
lists there.

Activate checkstyle in the View Model Tests
project to match reporting behaviour of the
build.  Fix existing (and new with this change
set) style problems.

Change-Id: I198c2df06b95463641c00d7f9bf47a4dbdbb0a7a
Signed-off-by: Christian W. Damus <give.a.damus@gmail.com>