Bug 543190 - [Validation Service] Reduce scope of re-validation for child context
Implement partial re-validation of the model on addition
of a child context, covering only the controls presented
by that child context and its containing views.

Includes introduction of a new abstract class with an
attribute in the Ecore model used for validation tests to
implement a realistic view model for the test scenario,
which implies re-generation of that Ecore model.

To simplify debugging of process of UniqueSettings,
implement toString() for that class using a generally
reusable new Labelizer utility.

Change-Id: I61eb76e847f5972b01eab9faed06730b393596bb
Signed-off-by: Christian W. Damus <give.a.damus@gmail.com>