Bug 544499: [Builder] Generalize the View Model Builder for Extensible Validation
Generalize the View Model Nature and Builder as an extensible Validation
Nature and Builder, with protocols for OSGi services to contribute
  - validation delegates performing the validation of a file
  - marker utility providers to customize, if necessary, marker creation

Refactor the view model builder as a specialization of the validation
builder that only supports view model files (not all contributed
delegate services).  In the case that a project also has the Validation
Nature, the View Model builder does nothing because the other builder
covers it.  This then supports both legacy projects (already having
the view model nature) and users that only want view model validation.

Implement validation delegates in builder bundles for each of the
editors provided by the EMF Forms IDE Tooling feature:  Ecore,
View Template, and Data Template.  To ensure correct identification
of files to be validated, generate content type recognition for
view template and data template models.

Change-Id: I673a5c9f437e3fb88577a03a4560bd646e4ae8b6
Signed-off-by: Christian W. Damus <give.a.damus@gmail.com>