Bug 542671 - [Segments] Adapt the view template's DMR selectors to segments
* Add a root EClass to the DomainModelReferenceSelector
* Move SelectEClassWizardPage to editor.controls bundle
  * Remove checkbox asking for control generation
  * Add subclass SelectEClassForViewWizardPage containing the checkbox
* Move SelectEcorePage to editor.controls bundle
* The CreateSegmentDmrWizard now allows to select a root EClass if none
is given. For this the SelectEcorePage and SelectEClassWizardPage are
reused.
* Add DmrSelectorSegmentDmrControlSWTRenderer to create segment based
DMRs for dmr selectors. When creating a dmr, this renderer also sets the
selector's root EClass
* In the Template Editor's init method, call GenericEditor.init at the
end to make sure the template model is migrated and referenced Ecores
registered before it is finally loaded. Also resolve all proxies before
showing the editor.

Note: The re-generated classes of the dmr selector package have lots of
changes because the code was not properly formatted before.

Change-Id: I213a76abfcd58c3af3135f8f4c84e5bc65fec2dd
Signed-off-by: Lucas Koehler <lkoehler@eclipsesource.com>