Bug 545460 - Generalize copy/paste support in master-detail tree editors
Implement a master/detail selection provider that encapsulates both
the master tree selection and the detail selection (currently always
empty) to ensure that the EMF implementations of the cut/copy/paste
actions provided by the action-bar contributor are disabled when the
user is working in text fields in the detail pane.  Employ this new
selection provider in the GenericEditor's TreeMasterDetailComposite and
in the TreeMasterDetailSWTRenderer to cover the two most common
implementations of form editors with master/detail trees.

Also fix the ViewEditorPart_PTest so that the tests that simulate the
keyboard work on Mac (Cocoa) platform.

Change-Id: I2568e56a4df8a06834c0a5322b6a43ca270dd93a
Signed-off-by: Christian W. Damus <give.a.damus@gmail.com>