Bug 546158 - [Segments] Extend IsProxyCondition tooling for segments
* DMRs for IsProxyConditions can no longer be created in the tree
(removed the child descriptor) but can now be created in the detail
view.
* Add general reference service strategies for rule condition dmrs
* Added handling for nested conditions: The root EClass of a condition's
dmr is now determined by walking up the containment hierarchy and
resolving the root EClass top to bottom (top == VView.rootEClass)
* The root EClass determination is extracted in a reusable util
* Added a new rule condition domain model reference renderer applicable
for rendering DMRs contained in a condition. It uses the aforementioned
util to resolve a condition dmr's root EClass. This allows the correct
display of the DMR's label for nested conditions.
* The reference service strategies for the leaf condition's dmr now
extends the new strategies. Only the feature selection validation and
the handles method are overridden


Tests:
Added test cases for the new reference service strategies and the root
EClass util.

Change-Id: Id637d9c9d518e88af8268c13788536255ed23be7
Signed-off-by: Lucas Koehler <lkoehler@eclipsesource.com>