Bug 530314 - Shall we keep AbstractFeatureAwareBuilder?
Deprecate the AbstractFeatureAwareBuilder. Add specific
fluent methods to its subclasses to enable/disable the
features.  Use the features only in the configuration
to communicate settings to the controls internally.

Also, a previous fix for the reliance on the current mouse
pointer location to guess the column that the user is
currently interacting with was incomplete, so this problem
is here addressed also in the abstract GridColumnAction.

Change-Id: I56c2632c3a8d29e4931774df287b4a27e481a210
Signed-off-by: Christian W. Damus <give.a.damus@gmail.com>