Bug 545686: Keyboard navigation of tables with many validation problems is slow
Avoid costly calls to the default contains(...) implementation
in the diagnostic list for diagnostics that would have to be
cached in the map (which is more efficient to search).

Also add a work-around for failure of RCPTT tests (which
do not make real mouse events) to track the mouse pointer
in the Grid Table Renderer for calculation of the available
column actions.

Change-Id: Ie2f9066f1f84a055055e929d5c803cedb9ca3acc
Signed-off-by: Christian W. Damus <give.a.damus@gmail.com>