Bug 547271 - [Performance] Filtering and adding objects to large tables is slow
Provide a more efficient filtering mechanism and avoid redundant
viewer refreshes.

Cache a generated view model for each EClass that needs it and clone
in the usual way (plus tweaks for read-only controls) for each object
of that class.

Don’t set properties of cells that don’t need updating, because it is
expensive to do so (especially fg/bg colours and text).

Change-Id: I47a3fd894e3e692252fc4448806dc827ad2ab671
Signed-off-by: Christian W. Damus <give.a.damus@gmail.com>