Bug 547504: ViewModelHelper should not escalate
The ViewModelHelper now just logs a warning if it can't find a
referenced Ecore of a given view.

The ViewModelHelper is used to load a view from a file. It checks the
referenced Ecore paths of a given view and tries to load these Ecores
from the current workspace. When it can't find a file in the given path
the ViewModelHelper previously threw a FileNotFoundException. This is
not strictly necessary and lead to errors when a view was for example
misconfigured.

Includes testcases.

Change-Id: I1fc53f1d0a41628a1aba16f8a10deb7e63fd3111
Signed-off-by: Stefan Dirix <sdirix@eclipsesource.com>